+++ src/bedrock/core/utility/enable_non_owner_references.h
@@ -1,0 +1,23 @@
+#pragma once
+
+#include <memory>
+
+namespace Bedrock {
+
+class EnableNonOwnerReferences {
+public:
+    virtual ~EnableNonOwnerReferences() = default;
+
+    struct ControlBlock {
+        bool mIsValid{false};
+    };
+
+    [[nodiscard]] const std::shared_ptr<ControlBlock> &controlBlock() const { return mControlBlock; }
+
+private:
+    std::shared_ptr<ControlBlock> mControlBlock;
+};
+
+static_assert(sizeof(EnableNonOwnerReferences) == 24);
+
+}  // namespace Bedrock