+++ src/bedrock/network/packet.h
@@ -1,14 +1,20 @@
 #pragma once
 
+#include <cstddef>
+#include <optional>
+#include <string>
 #include <string_view>
 
 #include "bedrock/common_types.h"
 #include "bedrock/core/utility/binary_stream.h"
 #include "bedrock/network/minecraft_packet_ids.h"
+#include "bedrock/network/serialization_mode.h"
 #include "bedrock/platform/result.h"
 
+class BinaryStream;
+
 namespace cereal {
-class ReflectionCtx;
+struct ReflectionCtx;
 }
 
 class Packet {
@@ -16,7 +22,24 @@ public:
     virtual ~Packet() = default;
     [[nodiscard]] virtual MinecraftPacketIds getId() const = 0;
     [[nodiscard]] virtual std::string_view getName() const = 0;
+    [[nodiscard]] virtual std::size_t getMaxSize() const;
+    [[nodiscard]] virtual Bedrock::Result<void> checkSize(std::size_t packet_size, bool receiver_is_server) const;
+    virtual void writeWithSerializationMode(BinaryStream &stream, const cereal::ReflectionCtx &reflection_ctx,
+                                            std::optional<SerializationMode> override_mode) const;
+    virtual void write(BinaryStream &stream, const cereal::ReflectionCtx &reflection_ctx) const;
+    virtual void write(BinaryStream &stream) const = 0;
+    virtual Bedrock::Result<void> read(ReadOnlyBinaryStream &stream, const cereal::ReflectionCtx &reflection_ctx);
+    virtual Bedrock::Result<void> read(ReadOnlyBinaryStream &stream);
+    [[nodiscard]] virtual bool disallowBatching() const;
+    [[nodiscard]] virtual bool isValid() const;
+    [[nodiscard]] virtual SerializationMode getSerializationMode() const;
+    virtual void setSerializationMode(SerializationMode mode);
+    [[nodiscard]] virtual std::string toString() const;
 
     Bedrock::Result<void> readNoHeader(ReadOnlyBinaryStream &stream, const cereal::ReflectionCtx &reflection_ctx,
                                        const SubClientId &sub_id);
+
+private:
+    virtual Bedrock::Result<void> _read(ReadOnlyBinaryStream &stream, const cereal::ReflectionCtx &reflection_ctx);
+    virtual Bedrock::Result<void> _read(ReadOnlyBinaryStream &stream) = 0;
 };